Initial Workings
Maintain translated content for posts: Every public post_type has a
shadow post_type for each of the site languages. In time these shadow
post_types will not have a public UI in the admin area. Swap content by
hooking parse_request to target the appropriate shadow post type.

Link each translated post to the source post in the default language:
Each post/page/whatever which is not in the default language can be
linked back to it's original version in the default language. Currently
this process is manual; you type in a post ID in a metabox, in time
this will be maintained by the plugin.

Fallback to content in default language where translations are not
available: When a post is shown in a language which is not the default
language, if the title, content or excerpt for that post is not
available in that language the title/content/excerpt from the default
language is substituted.