Language switching in admin area
Adds a menu to the admin bar to select the language from the available
languages in the site. Some changes to the parse_request functionality
which allows us to switch between the translated shadow post_types
dependant on the current language.

Adds a taxonomy, post_translation, in preparation for moving the
linking of translated versions of a post in the default language into a
taxonomy rather than storing the relationship in post_meta.