Associate transid with new posts in a translation grouping
When a new post screen is passed a term_id as a GET param of
`sil_transid` that transid is associated with the new post, and the
post is set to the appropriate shadow post type (if not in the default
language). If the new post is in the default language, a new transid is
generated in the post_translation tax and associated with the post.
Closes #22.

Also renames SIL_DEFAULT_LANGUAGE to SIL_DEFAULT_LANG.