When editing, allow switching between translations
When in the admin area and editing, the admin bar now switches you
between translations. Closes #24.

When editing a post in admin, provide a "Create for %s" admin bar link
for missing translations.

New API function `sil_get_new_translation_url` which provides an admin
area URL to create a translation for a given language.