Adds a fairly simple WordPress widget offering language switching functionality.
User can choose to show available languages as a <UL> list or <SELECT> drop-down (using the same js as the core WP archives drop-down).

There's also a checkbox allowing you to show or exclude languages when there's no appropriate equivalent to link to. In list mode, it just adds a <LI> with no link. In drop-down mode, the <option> is disabled.