If the cache for IDs hits, then convert to post objects.
@simonwheatley, I typically do this rather than a `get_posts()` with
`include` call, as that's likely what `WP_Query` is going to do sinse
the introduction of `$split_the_query` anyway. And also, this means we
don't have to do the initial posts SELECT query either. However, in the
case of no persistant object cache, this can mean it's a little slower
(per $split_the_query again).

However, in the case of a Persistent Object Cache, it's likely[maybe]
tht all posts are already in the object cache, which means we don't have
to touch the database at all