Babble Test Case: Reset the post types during tearDown
Some tests reset posts to a state they would be in if they
had been initiated in the context of a non-canonical locale.
Creating posts in the context of a non-canonical locale changes
whether they are or are not excluded from search. Resetting
this state during tearDown prevents unexpected results from
one test due to the actions of previous tests.